PHP interpreters moved under the Tester\Runner\Interpreters namespace (BC break)
