+++ tests/bootstrap.php
@@ -1,10 +1,13 @@
 <?php
 
+use Tester\Runner\Interpreters;
+
+
 require __DIR__ . '/../src/bootstrap.php';
 require __DIR__ . '/../src/Runner/PhpInterpreter.php';
-require __DIR__ . '/../src/Runner/ZendPhpInterpreter.php';
-require __DIR__ . '/../src/Runner/ZendPhpDbgInterpreter.php';
-require __DIR__ . '/../src/Runner/HhvmPhpInterpreter.php';
+require __DIR__ . '/../src/Runner/Interpreters/ZendPhpInterpreter.php';
+require __DIR__ . '/../src/Runner/Interpreters/ZendPhpDbgInterpreter.php';
+require __DIR__ . '/../src/Runner/Interpreters/HhvmPhpInterpreter.php';
 
 
 date_default_timezone_set('Europe/Prague');
@@ -19,10 +22,10 @@ function test(\Closure $function)
 function createInterpreter()
 {
 	if (defined('HHVM_VERSION')) {
-		return new Tester\Runner\HhvmPhpInterpreter(PHP_BINARY);
+		return new Interpreters\HhvmPhpInterpreter(PHP_BINARY);
 	} elseif (defined('PHPDBG_VERSION')) {
-		return new Tester\Runner\ZendPhpDbgInterpreter(PHP_BINARY, ' -c ' . Tester\Helpers::escapeArg(php_ini_loaded_file()));
+		return new Interpreters\ZendPhpDbgInterpreter(PHP_BINARY, ' -c ' . Tester\Helpers::escapeArg(php_ini_loaded_file()));
 	} else {
-		return new Tester\Runner\ZendPhpInterpreter(PHP_BINARY, ' -c ' . Tester\Helpers::escapeArg(php_ini_loaded_file()));
+		return new Interpreters\ZendPhpInterpreter(PHP_BINARY, ' -c ' . Tester\Helpers::escapeArg(php_ini_loaded_file()));
 	}
 }
